Extract functionality to a handler to keep routes clean
