+++ support-backend/src/routers/apiRouter.ts
@@ -1,22 +1,6 @@
 import { Router } from 'express';
+import { postcodeLookupHandler } from '../handlers/postcodeLookup';
 
 export const apiRouter = Router();
 
-interface Address {
-	lineOne?: string;
-	lineTwo?: string;
-	city?: string;
-	state?: string;
-	postCode?: string;
-	country?: string;
-}
-
-apiRouter.get('/postcode-lookup/:postcode', (req, res) => {
-	const postcode = decodeURIComponent(req.params.postcode);
-	if (postcode.length > 10) {
-		res.status(400).send();
-		return;
-	}
-	const addresses: Address[] = [];
-	res.json(addresses);
-});
+apiRouter.get('/postcode-lookup/:postcode', postcodeLookupHandler);
